Refactor all "int" to "int32" for sizeof() compatibility. Fixes #683 (#687)

+++ noarch/time.go
@@ -53,15 +53,15 @@ func TimeTToFloat64(t TimeT) float64 {
 // Structure containing a calendar date and time broken down into its
 // components
 type Tm struct {
-	Tm_sec   int
-	Tm_min   int
-	Tm_hour  int
-	Tm_mday  int
-	Tm_mon   int
-	Tm_year  int
-	Tm_wday  int
-	Tm_yday  int
-	Tm_isdst int
+	Tm_sec   int32
+	Tm_min   int32
+	Tm_hour  int32
+	Tm_mday  int32
+	Tm_mon   int32
+	Tm_year  int32
+	Tm_wday  int32
+	Tm_yday  int32
+	Tm_isdst int32
 	// tm_gmtoff int32
 	// tm_zone   []byte
 }
@@ -72,14 +72,14 @@ type Tm struct {
 func LocalTime(timer []TimeT) (tm []Tm) {
 	t := time.Unix(int64(timer[0]), 0)
 	tm = make([]Tm, 1)
-	tm[0].Tm_sec = t.Second()
-	tm[0].Tm_min = t.Minute()
-	tm[0].Tm_hour = t.Hour()
-	tm[0].Tm_mday = t.Day()
-	tm[0].Tm_mon = int(t.Month()) - 1
-	tm[0].Tm_year = t.Year() - 1900
-	tm[0].Tm_wday = int(t.Weekday())
-	tm[0].Tm_yday = t.YearDay() - 1
+	tm[0].Tm_sec = int32(t.Second())
+	tm[0].Tm_min = int32(t.Minute())
+	tm[0].Tm_hour = int32(t.Hour())
+	tm[0].Tm_mday = int32(t.Day())
+	tm[0].Tm_mon = int32(t.Month() - 1)
+	tm[0].Tm_year = int32(t.Year() - 1900)
+	tm[0].Tm_wday = int32(t.Weekday())
+	tm[0].Tm_yday = int32(t.YearDay() - 1)
 	return
 }
 
@@ -88,14 +88,14 @@ func Gmtime(timer []TimeT) (tm []Tm) {
 	t := time.Unix(int64(timer[0]), 0)
 	t = t.UTC()
 	tm = make([]Tm, 1)
-	tm[0].Tm_sec = t.Second()
-	tm[0].Tm_min = t.Minute()
-	tm[0].Tm_hour = t.Hour()
-	tm[0].Tm_mday = t.Day()
-	tm[0].Tm_mon = int(t.Month()) - 1
-	tm[0].Tm_year = t.Year() - 1900
-	tm[0].Tm_wday = int(t.Weekday())
-	tm[0].Tm_yday = t.YearDay() - 1
+	tm[0].Tm_sec = int32(t.Second())
+	tm[0].Tm_min = int32(t.Minute())
+	tm[0].Tm_hour = int32(t.Hour())
+	tm[0].Tm_mday = int32(t.Day())
+	tm[0].Tm_mon = int32(t.Month() - 1)
+	tm[0].Tm_year = int32(t.Year() - 1900)
+	tm[0].Tm_wday = int32(t.Weekday())
+	tm[0].Tm_yday = int32(t.YearDay() - 1)
 	return
 }
 
@@ -103,10 +103,10 @@ func Gmtime(timer []TimeT) (tm []Tm) {
 // Returns the value of type time_t that represents the local time described
 // by the tm structure pointed by timeptr (which may be modified).
 func Mktime(tm []Tm) TimeT {
-	t := time.Date(tm[0].Tm_year+1900, time.Month(tm[0].Tm_mon)+1, tm[0].Tm_mday,
-		tm[0].Tm_hour, tm[0].Tm_min, tm[0].Tm_sec, 0, time.Now().Location())
+	t := time.Date(int(tm[0].Tm_year+1900), time.Month(tm[0].Tm_mon)+1, int(tm[0].Tm_mday),
+		int(tm[0].Tm_hour), int(tm[0].Tm_min), int(tm[0].Tm_sec), 0, time.Now().Location())
 
-	tm[0].Tm_wday = int(t.Weekday())
+	tm[0].Tm_wday = int32(t.Weekday())
 
 	return TimeT(int32(t.Unix()))
 }
